Move gitcloner code into cli folder
