+++ internal/cmd/cli/root.go
@@ -4,10 +4,9 @@ package cli
 import (
 	"github.com/spf13/cobra"
 
-	gitcmd "github.com/rancher/fleet/cmd/gitcloner/cmd"
-	"github.com/rancher/fleet/cmd/gitcloner/gogit"
 	"github.com/rancher/fleet/internal/client"
 	command "github.com/rancher/fleet/internal/cmd"
+	"github.com/rancher/fleet/internal/cmd/cli/gitcloner"
 	"github.com/rancher/fleet/pkg/version"
 )
 
@@ -34,7 +33,7 @@ func App() *cobra.Command {
 
 		NewTarget(),
 		NewDeploy(),
-		gitcmd.New(gogit.NewCloner()),
+		gitcloner.New(gitcloner.NewCloner()),
 	)
 
 	return root
